feat(input): 생성자 제거
